セレクターを分割
