+++ src/features/game/GameRegister.tsx
@@ -6,7 +6,7 @@ import { GameForm } from './components/GameForm';
 import { LeagueRule } from '../../types/league';
 import { GameFormData } from './types/form';
 import { useLoading } from '../../hooks';
-import { gamePlayerSelector } from './recoil/selectors/gameSelector';
+import { gamePlayerSelector } from './recoil/selectors';
 
 export const GameRegister = (props: {
   leagueID: string;
