+++ src/features/game/recoil/selectors/gameListSelector.ts
@@ -1,0 +1,11 @@
+import { selectorFamily } from 'recoil';
+import { fetchGameList } from '../../../../api/services/gameService';
+
+export const gameListSelector = selectorFamily({
+  key: 'gameListSelector',
+  get: (leagueId?: string) => async () => {
+    if (!leagueId) return null;
+    const res = await fetchGameList(leagueId);
+    return res;
+  },
+});
